Move GroupBy to Submodule and Add FutureWarning (#20506)
