+++ pandas/core/resample.py
@@ -7,9 +7,10 @@ from textwrap import dedent
 import pandas as pd
 from pandas.core.base import GroupByMixin
 
-from pandas.core.groupby import (BinGrouper, Grouper, _GroupBy, GroupBy,
-                                 SeriesGroupBy, groupby, PanelGroupBy,
-                                 _pipe_template)
+from pandas.core.groupby.groupby import (
+    BinGrouper, Grouper, _GroupBy, GroupBy, SeriesGroupBy, groupby,
+    PanelGroupBy, _pipe_template
+)
 
 from pandas.tseries.frequencies import to_offset, is_subperiod, is_superperiod
 from pandas.core.indexes.datetimes import DatetimeIndex, date_range
