+++ pandas/core/reshape/pivot.py
@@ -6,7 +6,7 @@ from pandas.core.dtypes.generic import ABCDataFrame, ABCSeries
 
 from pandas.core.reshape.concat import concat
 from pandas.core.series import Series
-from pandas.core.groupby import Grouper
+from pandas.core.groupby.groupby import Grouper
 from pandas.core.reshape.util import cartesian_product
 from pandas.core.index import Index, _get_objs_combined_axis
 from pandas.compat import range, lrange, zip
