+++ pandas/tests/groupby/test_timegrouper.py
@@ -57,11 +57,12 @@ class TestGroupBy(object):
             result3 = df.groupby(pd.Grouper(freq='5D')).sum()
             assert_frame_equal(result3, expected)
 
-    def test_groupby_with_timegrouper_methods(self):
+    @pytest.mark.parametrize("should_sort", [True, False])
+    def test_groupby_with_timegrouper_methods(self, should_sort):
         # GH 3881
         # make sure API of timegrouper conforms
 
-        df_original = pd.DataFrame({
+        df = pd.DataFrame({
             'Branch': 'A A A A A B'.split(),
             'Buyer': 'Carl Mark Carl Joe Joe Carl'.split(),
             'Quantity': [1, 3, 5, 8, 9, 3],
@@ -75,16 +76,18 @@ class TestGroupBy(object):
             ]
         })
 
-        df_sorted = df_original.sort_values(by='Quantity', ascending=False)
+        if should_sort:
+            df = df.sort_values(by='Quantity', ascending=False)
 
-        for df in [df_original, df_sorted]:
-            df = df.set_index('Date', drop=False)
-            g = df.groupby(pd.Grouper(freq='6M'))
-            assert g.group_keys
-            assert isinstance(g.grouper, pd.core.groupby.BinGrouper)
-            groups = g.groups
-            assert isinstance(groups, dict)
-            assert len(groups) == 3
+        df = df.set_index('Date', drop=False)
+        g = df.groupby(pd.Grouper(freq='6M'))
+        assert g.group_keys
+
+        import pandas.core.groupby.groupby
+        assert isinstance(g.grouper, pandas.core.groupby.groupby.BinGrouper)
+        groups = g.groups
+        assert isinstance(groups, dict)
+        assert len(groups) == 3
 
     def test_timegrouper_with_reg_groups(self):
